time deposits api fixed and ui

+++ server/src/main/java/com/upo/ebank/model/enums/DepositType.java
@@ -1,17 +1,15 @@
 package com.upo.ebank.model.enums;
 
-import java.math.BigDecimal;
-
 public enum DepositType {
-    ONE_MONTH(1, BigDecimal.valueOf(0.01)),
-    SIX_MONTHS(6, BigDecimal.valueOf(0.03)),
-    TWELVE_MONTHS(12, BigDecimal.valueOf(0.06)),
-    TWENTY_FOUR_MONTHS(24, BigDecimal.valueOf(0.08));
+    ONE_MONTH(1, 0.01),
+    SIX_MONTHS(6, 0.03),
+    TWELVE_MONTHS(12, 0.06),
+    TWENTY_FOUR_MONTHS(24, 0.08);
 
     private final int months;
-    private final BigDecimal interestRate;
+    private final double interestRate;
 
-    DepositType(int months, BigDecimal interestRate) {
+    DepositType(int months, double interestRate) {
         this.months = months;
         this.interestRate = interestRate;
     }
@@ -20,7 +18,7 @@ public enum DepositType {
         return months;
     }
 
-    public BigDecimal getInterestRate() {
+    public double getInterestRate() {
         return interestRate;
     }
 }
